+++ drools-compiler/src/test/java/org/drools/compiler/integrationtests/SharingTest.java
@@ -1,5 +1,8 @@
 package org.drools.compiler.integrationtests;
 
+import java.util.List;
+import java.util.Map;
+
 import org.drools.compiler.Cheese;
 import org.drools.compiler.CommonTestMethodBase;
 import org.drools.compiler.Person;
@@ -17,10 +20,10 @@ import org.drools.core.reteoo.builder.MethodCountingNodeFactory;
 import org.drools.core.reteoo.builder.NodeFactory;
 import org.junit.Test;
 import org.kie.api.definition.rule.Rule;
+import org.kie.api.io.ResourceType;
+import org.kie.api.runtime.KieSession;
 import org.kie.internal.KnowledgeBase;
-
-import java.util.List;
-import java.util.Map;
+import org.kie.internal.utils.KieHelper;
 
 import static org.drools.core.util.DroolsTestUtil.rulestoMap;
 
@@ -185,5 +188,76 @@ public class SharingTest extends CommonTestMethodBase {
         MethodCountingObjectTypeNode betaOTN  = (MethodCountingObjectTypeNode) joinNode.getRightInput();
         Map<String, Integer> countingMap = betaOTN.getMethodCountMap();
         assertNull(countingMap);
+    }
+
+    public static class TestStaticUtils {
+        public static int return1() {
+            return 1;
+        }
+    }
+
+    @Test
+    public void testShouldAlphaShareBecauseSameConstantDespiteDifferentSyntax() {
+        // DROOLS-1404
+        String drl1 = "package c;\n" +
+                      "import " + Misc2Test.TestObject.class.getCanonicalName() + "\n" +
+                      "rule fileArule1 when\n" +
+                      "  TestObject(value == 1)\n" +
+                      "then\n" +
+                      "end\n" +
+                      "";
+        String drl2 = "package iTzXzx;\n" + // <<- keep the different package
+                      "import " + Misc2Test.TestObject.class.getCanonicalName() + "\n" +
+                      "import " + TestStaticUtils.class.getCanonicalName() + "\n" +
+                      "rule fileBrule1 when\n" +
+                      "  TestObject(value == TestStaticUtils.return1() )\n" +
+                      "then\n" +
+                      "end\n" +
+                      "rule fileBrule2 when\n" + // <<- keep this rule
+                      "  TestObject(value == 0 )\n" +
+                      "then\n" +
+                      "end\n" +
+                      "";
+
+        KieSession kieSession = new KieHelper()
+                .addContent( drl1, ResourceType.DRL )
+                .addContent(drl2, ResourceType.DRL)
+                .build().newKieSession();
+
+        kieSession.insert(new Misc2Test.TestObject( 1) );
+
+        assertEquals(2, kieSession.fireAllRules() );
+    }
+
+    @Test
+    public void testShouldAlphaShareNotEqualsInDifferentPackages() {
+        // DROOLS-1404
+        String drl1 = "package c;\n" +
+                      "import " + Misc2Test.TestObject.class.getCanonicalName() + "\n" +
+                      "rule fileArule1 when\n" +
+                      "  TestObject(value >= 1 )\n" +
+                      "then\n" +
+                      "end\n" +
+                      "";
+        String drl2 = "package iTzXzx;\n" + // <<- keep the different package
+                      "import " + Misc2Test.TestObject.class.getCanonicalName() + "\n" +
+                      "rule fileBrule1 when\n" +
+                      "  TestObject(value >= 1 )\n" +
+                      "then\n" +
+                      "end\n" +
+                      "rule fileBrule2 when\n" + // <<- keep this rule
+                      "  TestObject(value >= 2 )\n" +
+                      "then\n" +
+                      "end\n" +
+                      "";
+
+        KieSession kieSession = new KieHelper()
+                .addContent(drl1, ResourceType.DRL)
+                .addContent(drl2, ResourceType.DRL)
+                .build().newKieSession();
+
+        kieSession.insert(new Misc2Test.TestObject( 1) );
+
+        assertEquals(2, kieSession.fireAllRules() );
     }
 }
