+++ drools-core/src/main/java/org/drools/core/rule/constraint/MvelConstraint.java
@@ -15,6 +15,20 @@
 
 package org.drools.core.rule.constraint;
 
+import java.io.Externalizable;
+import java.io.IOException;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.Executor;
+import java.util.concurrent.atomic.AtomicInteger;
+
 import org.drools.core.base.ClassFieldReader;
 import org.drools.core.base.DroolsQuery;
 import org.drools.core.base.EvaluatorWrapper;
@@ -56,20 +70,6 @@ import org.mvel2.compiler.CompiledExpression;
 import org.mvel2.compiler.ExecutableStatement;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.io.Externalizable;
-import java.io.IOException;
-import java.io.ObjectInput;
-import java.io.ObjectOutput;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.Executor;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.drools.core.reteoo.PropertySpecificUtil.*;
 import static org.drools.core.util.ClassUtils.areNullSafeEquals;
@@ -618,7 +618,7 @@ public class MvelConstraint extends MutableTypeConstraint implements IndexableCo
         Map<String, Object> otherImports = ((MVELDialectRuntimeData) kbase.getPackage( otherPkg ).getDialectRuntimeRegistry().getDialectData("mvel")).getImports();
 
         if (fieldValue != null) {
-            return equalsExpressionTokensInBothImports(getLeftInExpression(IndexUtil.ConstraintType.EQUAL), thisImports, otherImports);
+            return equalsExpressionTokensInBothImports(getLeftInExpression(constraintType), thisImports, otherImports);
         } else {
             return equalsExpressionTokensInBothImports(expression, thisImports, otherImports);
         }
